vulkan project setup
https://vulkan.lunarg.com/sdk/home#windows
https://drive.google.com/file/d/18B5RVOrgrSmq9GHDDvIBb4osN5wBc_Lq/view?usp=sharing
C:\Program Files (x86)\Microsoft Visual Studio

so what you'll need to run either of the two vulkan projects is
to install vulkan sdk 1.2.176.1 (which is the latest and the one I'm using) and extract the Installer zip file in the path provided.

not installing the correct version or extracting the Installer folder somewhere else will prevent compilation because the reference paths are hard-mapped. it's possible to get the project to run, but is tedious.

also, the project runs in 64 bit mode only; x86 fails to compile.